oss_helper/requirements.txt
oss2~=2.16.0
PyYAML~=6.0.1
requests~=2.28.1